Removed exception class org.jpmml.evaluator.MissingResultException
It is possible that the prediction logic cannot produce a result. For
example, this can happen if an input value is missing and the prediction
logic does not include a method for treating missing values.
Previous versions of the JPMML-Evaluator library threw an instance of
MissingResultException in such a situation. However, this behaviour did
not conform to the PMML specification, which requires that a missing
result should be handled gracefully by the Targets element.