Optimized the evaluation of 'if' built-in function
The boolean value of the first argument (IF clause) determines whether
the second argument (THEN part) or the third argument (ELSE part) is
returned as the result. In either case, one of the two arguments goes
unused.

It is possible to leave the unused argument untreated. This saves
evaluation time and guards against unexpected side effects.