Reverted commit fde2e35797
The pluginManagement mechanism is advertised as a means to configure
plugin information in the parent POM so that it could be referenced in
the children.
Apparently, this mechanism cannot be used for the "selective" execution
of Maven Deploy Plugin. Maven versions 3.0(.5) and 3.3(.3) execute it
automatically for all children, which results in empty distribution.