Fixed the evaluation of binary logistic regression models
Binary classification is encoded using two RegressionTable elements.
The first regression table, which has one or more predictor terms,
calculates the probability of the event taking place (p_{event}).
The second regression table, which does not have any predictor terms,
is not involved in probability calculations in any way.
The probability of the event not taking place (p_{no event}) is
calculated as (1.0 - p_{event}).