Fixed the inheritance of derived field values
DerivedField elements that are defined in the LocalTransformation
element of a model are "invisible" to its child models.
If a child model wants to "see" the value of a local derived field
ofits parent model, then it must be explicitly imported by
declaring an appropriate MiningField element.

Previous versions of the JPMML-Evaluator library wrongly assumed that a
child model must import the declaration of a derived field (ie. the
DerivedField element itself) and obtain its value by evaluating it in
the current evaluation context.