Avoiding unnecessary Cache and LoadingCache accesses
An Evaluator instance should maintain a private reference to the
cached data structure. This reference should be transient. The idea
is that if the Evaluator instance is transferred to another JVM
(eg. using Java serialization API), then it should bring about the
initialization of a local data structure (as opposed to copying over
the remote data structure).