Optimized the evaluation of MapValues elements
The order of lookup operations should be from the most discriminative
column to the least discriminative column.
Currently, it is assumed that the optimal ordering of columns is the
same as the ordering of FieldColumnPair child elements. Future versions
of the JPMML-Evaluator library may be able to (optionally-) estimate the
discriminative power of columns, and perform lookup operations in the
correct order.